pyproject.toml
[project]
name = "chronotag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Liudas Sodonis", email = "liudas.sodonis@fastloop.ai" }
]
requires-python = ">=3.14"
dependencies = [
    "google-cloud-logging>=3.12.1",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]
